requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
supabase>=2.27.2
openai>=1.0.0
openai-agents>=0.7.0
python-dotenv>=1.0.0
